% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abscorcalc.R
\name{abscorcalc}
\alias{abscorcalc}
\title{Calculate Absolute Correlations}
\usage{
abscorcalc(object, method = "spearman", abs = TRUE)
}
\arguments{
\item{object}{data frame}

\item{method}{method for correlation}

\item{abs}{absolute correlation if `TRUE`}
}
\value{
data frame with design columns, `trait` names and `value` of absolute correlations
}
\description{
Calculate Absolute Correlations
}
