% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biggplot}
\alias{biggplot}
\title{Biplot using ggplot2 via ordr package}
\usage{
biggplot(bip_pca, scale.factor = 2)
}
\arguments{
\item{bip_pca}{}

\item{scale.factor}{}
}
\value{
object of class ggplot2
}
\description{
Biplot using ggplot2 via ordr package
}
