% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_traits.R
\name{bind_traits}
\alias{bind_traits}
\title{Bind trait data together into unified objects.}
\usage{
bind_traits(datasets, dirname = ".")
}
\arguments{
\item{datasets}{vector of dataset names.}
}
\value{
side effects: save data in local RDS and CSV files
}
\description{
Bind trait data together into unified objects.
}
