% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot_data}
\alias{biplot_data}
\title{Prepare biplot data}
\usage{
biplot_data(
  dat,
  traits = unique(dat$trait),
  factors = c("strain", "animal", "sex", "condition"),
  orders = c("module", "kME", "p.value", "size")
)
}
\arguments{
\item{dat}{data frame with appropriate columns}

\item{traits}{trait names to include}

\item{factors}{design column names}
}
\value{
data frame
}
\description{
Prepare biplot data
}
