% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot_signal}
\alias{biplot_signal}
\alias{biplot_stat}
\alias{biplot_pca}
\alias{biggplot}
\alias{biplot_evidence}
\title{Prepare biplot signal data}
\usage{
biplot_signal(
  object,
  datatraits = unique(object$datatrait),
  orders = c("module", "kME", "p.value", "size")
)

biplot_stat(object, datatraits = unique(object$datatrait), orders = terms)

biplot_pca(
  bip,
  size = c("module", "kME", "p.value", "size"),
  strain = "NONE",
  threshold
)

biggplot(bip_pca, axes = 1:2, scale.factor = 2)

biplot_evidence(
  object,
  ordername,
  xlab,
  threshold,
  strain = "NONE",
  axes = 1:2,
  ...
)
}
\arguments{
\item{object}{data frame}

\item{datatraits}{`dataset: trait` names to include}

\item{orders}{column names used for ordering and plot sizes}

\item{bip}{object from `biplot_signal`}

\item{strain}{strain name to highlight in biplot}

\item{threshold}{vector of threshold values}

\item{bip_pca}{object of class bip_pca}

\item{axes}{axes to plot}

\item{scale.factor}{scale factor for arrows vs points}

\item{ordername}{name of order column}

\item{xlab}{x label}

\item{...}{additional paramaters ignored}

\item{transpose}{transpose for `princomp` if `TRUE`}
}
\value{
data frame

data frame

data frame of PCA components

object of class ggplot2

gg plot object
}
\description{
Prepare biplot signal data

Prepare biplot stats

Get PCA components from biplot data

Biplot using ggplot2 via ordr package

BiPlot for Evidence vs Spread
}
