% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot_signal}
\alias{biplot_signal}
\title{Prepare biplot signal data}
\usage{
biplot_signal(
  object,
  datatraits = unique(object$datatrait),
  orders = c("module", "kME", "p.value", "size")
)
}
\arguments{
\item{object}{data frame with appropriate columns}

\item{datatraits}{`dataset: trait` names to include}

\item{orders}{column names used for ordering and plot sizes}
}
\value{
data frame
}
\description{
Prepare biplot signal data
}
