% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot_stat}
\alias{biplot_stat}
\title{Prepare biplot stats}
\usage{
biplot_stat(object, datatraits = unique(object$datatrait), orders = terms)
}
\arguments{
\item{object}{data frame with appropriate columns}

\item{datatraits}{`dataset: trait` names to include}

\item{orders}{column names used for ordering and plot sizes}
}
\value{
data frame
}
\description{
Prepare biplot stats
}
