% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ind_signal.R
\name{calc_ind_signal}
\alias{calc_ind_signal}
\title{Calculate individual signal}
\usage{
calc_ind_signal(traitData, traitSignal)
}
\arguments{
\item{traitData}{data frame with harmonized data}

\item{traitSignal}{data frame with `signal` and `cellmean`}
}
\value{
data frame with `value` as individual signal, plus `signal` column
}
\description{
Calculate individual signal
}
