% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCcolors-data.R
\name{CCcolors}
\alias{CCcolors}
\title{Collaborative Cross colors}
\source{
<https://csbio.unc.edu/CCstatus/index.py?run=AvailableLines.information>
}
\description{
A vector of 8 colors for use with the mouse
Collaborative Cross and Diversity Outbreds.
}
\details{
`CCorigcolors` (see <https://github.com/rqtl/qtl2/blob/main/R/CCcolors-data.R>) are the original eight colors for the Collaborative Cross founder strains.
`CCcolors` are derived from the the Okabe-Ito color blind friendly palette in
 Wong (2011) Nature Methods \doi{10.1038/nmeth.1618}.
}
\examples{
data(CCcolors)
}
\keyword{datasets}
