% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{condition_biplot}
\alias{condition_biplot}
\title{BiPlot for Condition Contrasts}
\usage{
condition_biplot(object, ordername, xlab, threshold, strain = "NONE", ...)
}
\arguments{
\item{object}{data frame}

\item{ordername}{name of order column}

\item{xlab}{x label}

\item{threshold}{vector of threshold values}

\item{strain}{strain name to highlight in biplot}

\item{...}{additional paramaters ignored}
}
\value{
gg plot object
}
\description{
BiPlot for Condition Contrasts
}
