% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_compare.R
\name{cor_compare}
\alias{cor_compare}
\title{Compare correlation matrices using upper triangle}
\usage{
cor_compare(traitsum, cor1, cor2, ...)
}
\arguments{
\item{traitsum}{data frame with trait summaries}

\item{cor1, cor2}{correlation matrices to compare}

\item{...}{additional parameters (not used)}
}
\value{
tibble of data frame
}
\description{
Compare correlation matrices using upper triangle
}
