% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_compare.R
\name{cor_extreme}
\alias{cor_extreme}
\title{Extreme correlation comparison summaries}
\usage{
cor_extreme(
  traitsum,
  object = cor_compare(traitsum, ...),
  ...,
  cormin = 0.8,
  minlogp = 2
)
}
\arguments{
\item{traitsum}{data frame with trait summaries}

\item{object}{object from `cor_compare()`}

\item{...}{additional parameters}

\item{cormin}{minimum correlation to keep}

\item{minlogp}{minimul log of p-value to keep}
}
\value{
data frame
}
\description{
Extreme correlation comparison summaries
}
