% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trait_query_func.R
\name{create_trait_query_func}
\alias{create_trait_query_func}
\title{Create a function to query Trait Data}
\usage{
create_trait_query_func(dbfile = NULL, db = NULL, table_name = "traitData")
}
\arguments{
\item{dbfile}{Name of database file}

\item{db}{Optional database connection (provide one of `dbfile` and `db`).}

\item{table_name}{Name of table in the database}
}
\value{
Function with two arguments, `dataset` and `trait`,
    which returns a data frame with the traits.

This requires an existing SQLite database of traits with names at least
    `traitData`, `traitStats`, `traitSignal`.
}
\description{
Create a function that will connect to a SQLite database of founder
variant information and return a data frame with variants for a
selected region.
}
