% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trait_query_func.R
\name{create_trait_sqlite}
\alias{create_trait_sqlite}
\title{Create Trait SQLite File}
\usage{
create_trait_sqlite(traitData, filename = "traitData.sqlite")
}
\arguments{
\item{traitData}{data frame}

\item{filename}{name of SQLite file to create}
}
\value{
list of tables (invisbly)
}
\description{
Create Trait SQLite File
}
