% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectplot.R
\name{effectplot}
\alias{effectplot}
\title{Effect Plots}
\usage{
effectplot(object, traitnames = udatatraits)
}
\arguments{
\item{object}{data frame from `strainstats()`}

\item{traitnames}{`traitnames` to show}
}
\description{
Effect Plots
}
