% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{foundrArrange}
\alias{foundrArrange}
\title{Arrange Founder Traits in Order}
\usage{
foundrArrange(traitsumdata, order, datatypes = "")
}
\arguments{
\item{traitsumdata}{data frame with summary}

\item{order}{name to use to order traits}

\item{datatypes}{type to subset `traitsumdata`}
}
\value{
data frame reordered
}
\description{
Arrange Founder Traits in Order
}
