% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{foundrDownloads}
\alias{foundrDownloads}
\title{Downloads for Founder App}
\usage{
foundrDownloads(trait, datatype)
}
\arguments{
\item{trait}{trait name(s)}

\item{datatype}{type of data}
}
\value{
uses shiny downloadHandler
}
\description{
See `foundrServer` for details of download outputs.
}
