% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{foundrSettings}
\alias{foundrSettings}
\title{Settings for Founder Traits}
\usage{
foundrSettings(traitsumdata)
}
\arguments{
\item{traitsumdata}{data frame with summary}
}
\value{
`renderUI` object
}
\description{
Settings for Founder Traits
}
