% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{foundrUI}
\alias{foundrUI}
\title{User Interface for Founder Shiny App}
\usage{
foundrUI(title)
}
\arguments{
\item{title}{title of shiny app}
}
\value{
A UI definition that can be passed to the `shinyUI` function.
}
\description{
See foundrServer for needed user-supplied functions.
}
