% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_spread_evidence.R
\name{ggplot_evidence_spread}
\alias{ggplot_evidence_spread}
\title{GGplot of evidence vs spread}
\usage{
ggplot_evidence_spread(
  object,
  xlab = NULL,
  plottype = c("dotplot", "volcano", "biplot"),
  ordername = attr(object, "ordername"),
  ntraits = 20,
  ...
)
}
\arguments{
\item{object}{data frame with optional attributes}

\item{xlab}{replacement horizontal label if not `NULL`}

\item{plottype}{}

\item{ordername}{column name to order entries}

\item{ntraits}{number of traits to plot}

\item{...}{additional parameters ignored}
}
\value{
GG plot object
}
\description{
This includes volcano plot and other plots that compare data spread to
strength of evidence. The strength of evidence is often a `p.value`
(typically plotted as `-log10(p.value)`) but need not be.
The spread may be differences of mean values or
SDs (square root of MS for model term).
}
