% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_compare.R
\name{ggplot_extreme}
\alias{ggplot_extreme}
\title{GGplot of extreme correlation data}
\usage{
ggplot_extreme(
  object,
  summary_type,
  dat_extreme = cor_extreme(...)[1, ],
  ...,
  condition = ""
)
}
\arguments{
\item{object}{data frame with appropriate data summaries}

\item{summary_type}{type of summary to extract from `object`}

\item{dat_extreme}{data frame with extreme correlation summaries}

\item{...}{additional parameters}

\item{condition}{factors in form `"a + b"` to use for option faceting}
}
\value{
ggplot2 object
}
\description{
GGplot of extreme correlation data
}
