% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_module.R
\name{ggplot_module}
\alias{ggplot_module}
\title{GGplot of Module P-values}
\usage{
ggplot_module(object, terms = c("p_strain_sex", "p_strain"))
}
\arguments{
\item{object}{data frame with appropriate columns}

\item{terms}{names of two columns with p-values}
}
\value{
ggplot2 object
}
\description{
GGplot of Module P-values
}
