% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_strain_cond.R
\name{ggplot_strain_cond}
\alias{ggplot_strain_cond}
\title{GGplot by strain and condition}
\usage{
ggplot_strain_cond(
  datatraits,
  facet_strain = FALSE,
  condition = "sex_diet",
  boxplot = FALSE
)
}
\arguments{
\item{datatraits}{data frame to be plotted}

\item{facet_strain}{facet by strain if `TRUE`}

\item{condition}{name of column with condition}

\item{boxplot}{overlay boxplot if `TRUE`}
}
\value{
object of class ggplot
}
\description{
GGplot by strain and condition
}
