% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_time.R
\name{ggplot_strain_time}
\alias{ggplot_strain_time}
\title{GGplot of Strains over Time}
\usage{
ggplot_strain_time(object, ..., drop_xlab = TRUE, legend_position = "bottom")
}
\arguments{
\item{object}{object of class `strain_time`}

\item{...}{additional parameters}

\item{drop_xlab}{drop xlab for all but last plot if `TRUE`}

\item{legend_position}{position of legend ("none" for none)}
}
\value{
ggplot object
}
\description{
GGplot of Strains over Time
}
