% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_traitData.R
\name{ggplot_traitData}
\alias{ggplot_traitData}
\alias{strainplot}
\alias{autoplot.traitData}
\alias{plot.traitData}
\title{GGplot by strain and condition}
\usage{
ggplot_traitData(object, ...)

strainplot(object, ...)

\method{autoplot}{traitData}(object, ...)

\method{plot}{traitData}(object, ...)
}
\arguments{
\item{object}{data frame to be plotted}

\item{...}{additional parameters (ignored)}

\item{facet_strain}{facet by strain if `TRUE`}

\item{shape_sex}{use different shape by sex if `TRUE`}

\item{boxplot}{overlay boxplot if `TRUE`}

\item{horizontal}{flip vertical and horizontal axis if `TRUE`}
}
\value{
object of class ggplot
}
\description{
GGplot by strain and condition
}
\examples{
ggplot_traitData(sampleData)

}
