% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_traitSolos.R
\name{ggplot_traitSolos}
\alias{ggplot_traitSolos}
\alias{strainplot}
\alias{autoplot.traitSolos}
\alias{plot.traitSolos}
\title{GGplot by strain and condition}
\usage{
ggplot_traitSolos(object, ...)

strainplot(object, ...)

\method{autoplot}{traitSolos}(object, ...)

\method{plot}{traitSolos}(object, ...)
}
\arguments{
\item{object}{data frame to be plotted}

\item{...}{additional parameters (ignored)}

\item{facet_strain}{facet by strain if `TRUE`}

\item{shape_sex}{use different shape by sex if `TRUE`}

\item{boxplot}{overlay boxplot if `TRUE`}

\item{horizontal}{flip vertical and horizontal axis if `TRUE`}
}
\value{
object of class ggplot
}
\description{
GGplot by strain and condition
}
\examples{
sampleSignal <- partition(sampleData)
sampleSolo <- traitSolos(sampleData, sampleSignal)
ggplot_traitSolos(sampleSolo)

}
