% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize}
\alias{harmonize}
\title{Harmonize Dataset}
\usage{
harmonize(dataset, links, userHarmony, ...)
}
\arguments{
\item{dataset}{name of dataset}

\item{links}{data frame with `shortname`, `address`, `longname`}

\item{userHarmony}{function provided by user}

\item{...}{additional optional parameters to `userHarmony()`}
}
\value{
side action to save RDS files locally
}
\description{
Harmonize Dataset
}
