% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_signal.R
\name{join_signal}
\alias{join_signal}
\title{Join trait data and signal}
\usage{
join_signal(traitData, traitSignal, response = c("rest", "noise"))
}
\arguments{
\item{traitData}{data frame with harmonized data}

\item{traitSignal}{data frame with `signal` and `cellmean`}

\item{response}{name of response to create through joining}
}
\value{
data frame with `value` as individual signal, plus `signal` column
}
\description{
Join trait data and signal
}
