% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkpath.R
\name{linkpath}
\alias{linkpath}
\title{Title}
\usage{
linkpath(dataset, links, datadir = "data")
}
\arguments{
\item{dataset}{name of dataset desired}

\item{links}{data frame with links to files}

\item{datadir}{data directory (default is "data")}
}
\description{
Title
}
