% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleHarmony.R
\name{moduleHarmony}
\alias{moduleHarmony}
\title{Harmonize module object}
\usage{
moduleHarmony(object, response = names(object))
}
\arguments{
\item{object}{object of class traitModule}

\item{response}{name of response to extract}
}
\value{
harmonize object
}
\description{
Harmonize module object
}
