% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboxplot.R
\name{myboxplot}
\alias{myboxplot}
\title{Boxplot overlaid by jittered data}
\usage{
myboxplot(data, x)
}
\arguments{
\item{data}{data frame}

\item{x}{title}
}
\description{
Boxplot overlaid by jittered data
}
