% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderTraitStats.R
\name{orderTraitStats}
\alias{orderTraitStats}
\title{Order Trait Statistics}
\usage{
orderTraitStats(orders, traitStats)
}
\arguments{
\item{orders}{name of order criterion}

\item{traitStats}{data frame with statistics}
}
\value{
data frame
}
\description{
Order Trait Statistics
}
