% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition Trait into Signal, Rest and Noise}
\usage{
partition(
  object,
  trait = "trait",
  value = "value",
  signal = ifelse(is_condition, "strain * sex * condition", "strain * sex"),
  rest = ifelse(is_condition, "strain * sex + sex * condition", "sex")
)
}
\arguments{
\item{object}{data frame in long format with trait data}

\item{trait}{name of column with trait names}

\item{value}{name column with trait values}

\item{signal}{signal factor combination as string for `formula`}

\item{rest}{rest factor combination as string for `formula`}
}
\value{
data frame with added columns `signal`, `cellmean`
}
\description{
Partition a trait value into three parts based on rest and signal factors.
The signal is the main interest for comparison across traits, while the rest
concerns aspects of the experiment that are controlled. Noise is unexplained variation.
The three parts will be orthogonal to each other.
The `signal` factors, entered as terms for `formula`, are conditioned by the `rest` factors.
}
\examples{
partition(sampleData)
}
