% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_pair.R
\name{pivot_pair}
\alias{pivot_pair}
\title{Pivot pair of traits}
\usage{
pivot_pair(object, pair)
}
\arguments{
\item{object}{data frame with specific columns}

\item{pair}{2-vector of trait names}
}
\value{
data frame
}
\description{
Pivot pair of traits
}
