% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residit.R
\name{residit}
\alias{residit}
\title{Partition Trait into Signal, Design and Noise}
\usage{
residit(
  object,
  trait = "trait",
  value = "value",
  signal = "strain * sex * diet",
  design = "strain * sex + sex * diet"
)
}
\arguments{
\item{object}{data frame in long format with trait data}

\item{trait}{name of column with trait names}

\item{value}{name column with trait values}

\item{signal}{signal factor combination as string for `formula`}

\item{design}{design factor combination as string for `formula`}
}
\description{
Partition a trait value into three parts based on design and signal factors.
The signal is the main interest for comparison across traits, while the design
concerns aspects of the experiment that are controlled. Noise is unexplained variation.
The three parts will be orthogonal to each other.
The `signal` factors, entered as terms for `formula`, are conditioned by the `design` factors.
}
