% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData-data.R
\name{sampleData}
\alias{sampleData}
\title{Sample Data for Package foundr}
\description{
A data frame with columns `datatype`, `strain`, `sex`, `animal`, `condition`, `trait`, `value`
}
\details{
`sampleData` is a minimal data example for the foundr package.
}
\examples{
data(sampleData)
}
\keyword{datasets}
