% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatplot.R
\name{scatplot}
\alias{scatplot}
\title{Scatter plot by trait values}
\usage{
scatplot(data, x, y)
}
\arguments{
\item{data}{data frame}

\item{x}{first trait}

\item{y}{second trait}
}
\description{
Scatter plot by trait values
}
