% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_data_pairs.R
\name{select_data_pairs}
\alias{select_data_pairs}
\title{Select data including Key Trait and Related Datasets}
\usage{
select_data_pairs(object, key_trait, rel_dataset = NULL)
}
\arguments{
\item{object}{data frame}

\item{key_trait}{name of key trait}

\item{rel_dataset}{name of related datasets}
}
\value{
data frame
}
\description{
Select data including Key Trait and Related Datasets
}
