% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSignal.R
\name{selectSignal}
\alias{selectSignal}
\title{Select Signal}
\usage{
selectSignal(
  traitSignal,
  traitnames,
  strains = names(CCcolors),
  response = c("mean", "signal")
)
}
\arguments{
\item{traitSignal}{data frame}

\item{traitnames}{names of traits to subset}

\item{strains}{names of strains to subset}

\item{response}{name of response}
}
\value{
data frame
}
\description{
Select Signal
}
