% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTrait.R
\name{selectTrait}
\alias{selectTrait}
\title{Select Traits and Strains for Response}
\usage{
selectTrait(
  traitData,
  traitSignal,
  traitnames = unite_datatraits(traitSignal),
  strains = names(CCcolors),
  response = c("individual", "mean", "signal", "ind_signal"),
  abbrev = TRUE,
  sep = ": "
)
}
\arguments{
\item{traitData}{data frame}

\item{traitSignal}{data frame}

\item{traitnames}{names of `dataset: trait` combinations to subset}

\item{strains}{names of strains to subset}

\item{response}{name of response to return}

\item{abbrev}{abbreviate names if `TRUE`}
}
\value{
data frame
}
\description{
Select Traits and Strains for Response
}
