% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_GTT.R
\name{separate_GTT}
\alias{separate_GTT}
\title{Separate GTT Traits with Minute and Week}
\usage{
separate_GTT(object, traits)
}
\arguments{
\item{object}{data frame}

\item{traits}{trait names}
}
\value{
data frame
}
\description{
Separate GTT Traits with Minute and Week
}
