% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_time.R
\name{separate_time}
\alias{separate_time}
\title{Separate Time Traits}
\usage{
separate_time(object, datatraits, timeunit = c("week", "minute"))
}
\arguments{
\item{object}{data frame}

\item{datatraits}{data frame with dataset and trait names}

\item{timeunit}{name of time unit}
}
\value{
data frame
}
\description{
Separate Time Traits
}
