% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_week.R
\name{separate_week}
\alias{separate_week}
\title{Separate Week from Trait Name}
\usage{
separate_week(object, traits)
}
\arguments{
\item{object}{data frame}

\item{traits}{trait names}
}
\value{
data frame
}
\description{
Separate Week from Trait Name
}
