% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyBiplot.R
\name{shinyBiplotInput}
\alias{shinyBiplotInput}
\alias{shinyBiplotOutput}
\alias{shinyBiplot}
\title{Shiny Module Input for Biplots}
\usage{
shinyBiplotInput(id)

shinyBiplotOutput(id)

shinyBiplot(
  id,
  call_par,
  sex_par,
  main_par,
  contrastTable,
  info,
  threshold,
  ord_selection
)
}
\arguments{
\item{id}{identifier}

\item{sex_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}

\item{modTitle}{character string title for section}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Biplots

Shiny Module Output for Biplots

Shiny Module Server for Contrast Plots
}
