% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastModule.R
\name{shinyContrastModuleInput}
\alias{shinyContrastModuleInput}
\alias{shinyContrastModuleOutput}
\alias{shinyContrastModule}
\title{Shiny Module Input for Modules of Contrasts}
\usage{
shinyContrastModuleInput(id)

shinyContrastModuleOutput(id)

shinyContrastModule(
  id,
  panel_par,
  main_par,
  traitModule,
  moduleContrast,
  traitContrast,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitModule}{static data frames}

\item{moduleContrast, traitContast}{reactive data frames}

\item{customSettings}{list of custom settings}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Modules of Contrasts

Shiny Module Output for Modules of Contrasts

Shiny Module Server for Modules of Contrasts
}
