% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastPanel.R
\name{shinyContrastPanelInput}
\alias{shinyContrastPanelInput}
\alias{shinyContrastPanelOutput}
\alias{shinyContrastPanel}
\title{Shiny Module Input for Trait Panel}
\usage{
shinyContrastPanelInput(id)

shinyContrastPanelOutput(id)

shinyContrastPanel(
  id,
  main_par,
  traitSignal,
  traitStats,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitSignal, traitStats}{static data frames}

\item{customSettings}{list of custom settings}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Trait Panel

Shiny Module Output for Contrast Panel

Shiny Module Server for Contrast Panel
}
