% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastPlot.R
\name{shinyContrastPlotOutput}
\alias{shinyContrastPlotOutput}
\alias{shinyContrastPlot}
\title{Shiny Module Output for Contrast Plots}
\usage{
shinyContrastPlotOutput(id)

shinyContrastPlot(
  id,
  panel_par,
  main_par,
  contrastTable,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier}

\item{panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}
}
\value{
nothing returned

reactive object
}
\description{
Shiny Module Output for Contrast Plots

Shiny Module Server for Contrast Plots
}
