% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastSex.R
\name{shinyContrastSexOutput}
\alias{shinyContrastSexOutput}
\alias{shinyContrastSex}
\title{Shiny Sex Output for Contrast Plots}
\usage{
shinyContrastSexOutput(id)

shinyContrastSex(id, panel_par, main_par, contrastTable, customSettings = NULL)
}
\arguments{
\item{id}{identifier}

\item{panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}
}
\value{
nothing returned

reactive object
}
\description{
Shiny Sex Output for Contrast Plots

Shiny Sex Server for Contrast Plots
}
