% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCorPlot.R
\name{shinyCorPlotOutput}
\alias{shinyCorPlotOutput}
\alias{shinyCorPlot}
\title{Shiny Module Output for Trait Correlations}
\usage{
shinyCorPlotOutput(id)

shinyCorPlot(id, panel_par, main_par, CorTable, customSettings = NULL)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive inputs from calling modules}

\item{CorTable}{reactive data frames}

\item{customSettings}{static list of settings}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

reactive object
}
\description{
Shiny Module Output for Trait Correlations

Shiny Module Server for Trait Stats
}
