% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCorrelation.R
\name{shinyCorrelationUI}
\alias{shinyCorrelationUI}
\alias{shinyCorrelation}
\title{Shiny Module UI for Correlation Plot}
\usage{
shinyCorrelationUI(id)

shinyCorrelation(input, output, session, main_par, traitSignal)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{traitSignal}{reactive objects from `foundrServer`}
}
\value{
nothing returned

reactive object for `shinyCorrelationUI`
}
\description{
Shiny Module UI for Correlation Plot

Shiny Module Server for Correlation Plots
}
