% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCorTable.R
\name{shinyCorTableUI}
\alias{shinyCorTableUI}
\alias{shinyCorTableOutput}
\alias{shinyCorTable}
\title{Shiny Module UI for Trait Correlations}
\usage{
shinyCorTableUI(id)

shinyCorTableOutput(id)

shinyCorTable(id, stats_par, key_trait, traitSignal)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{stats_par}{reactive inputs from calling modules}

\item{traitSignal}{reactive data frames}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module UI for Trait Correlations

Shiny Module UI for Trait Correlations

Shiny Module Server for Trait Stats
}
