% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyDownloads.R
\name{shinyDownloadsOutput}
\alias{shinyDownloadsOutput}
\alias{shinyDownloads}
\title{Shiny Module Output for Downloads}
\usage{
shinyDownloadsOutput(id)

shinyDownloads(id, prefix, main_par, postfix, plotObject, tableObject)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{prefix}{static prefix for filename}

\item{main_par}{input parameters from calling routine}

\item{postfix}{reactive postfix for filename}

\item{tableObject}{reactive table pring object}

\item{plotObjec}{reactive plot print object}
}
\value{
nothing returned

nothing
}
\description{
Shiny Module Output for Downloads

Shiny Module Server for Downloads
}
