% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyEffects.R
\name{shinyEffectsUI}
\alias{shinyEffectsUI}
\alias{shinyEffects}
\title{Shiny Module UI for Effects Plot}
\usage{
shinyEffectsUI(id)

shinyEffects(
  input,
  output,
  session,
  main_par,
  traitStatsSelectType,
  traitStatsArranged
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{traitStatsSelectType, traitStatsArranged}{reactive objects from `foundrServer`}
}
\value{
nothing returned

reactive object for `shinyEffectsUI`
}
\description{
Shiny Module UI for Effects Plot

Shiny Module Server for effects Plots
}
