% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModuleComp.R
\name{shinyModuleCompUI}
\alias{shinyModuleCompUI}
\alias{shinyModuleComp}
\title{Shiny Module UI for Dendro Plot}
\usage{
shinyModuleCompUI(id)

shinyModuleComp(input, output, session, main_par, traitModule)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{traitModule}{reactive object with list created by `listof_wgcnamodules`}
}
\value{
nothing returned

reactive object for `shinyModuleCompUI`
}
\description{
Shiny Module UI for Dendro Plot

Shiny Module Server for Dendro Plots
}
