% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModuleDendro.R
\name{shinyModuleDendroUI}
\alias{shinyModuleDendroUI}
\alias{shinyModuleDendro}
\title{Shiny Module UI for Dendro Plot}
\usage{
shinyModuleDendroUI(id)

shinyModuleDendro(input, output, session, module_par, main_par, traitModule)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{module_par, main_par}{reactive arguments from `foundrServer` and `shinyModules`}

\item{traitModule}{reactive object with list created by `listof_wgcnamodules`}
}
\value{
nothing returned

reactive object for `shinyModuleDendroUI`
}
\description{
Shiny Module UI for Dendro Plot

Shiny Module Server for Dendro Plots
}
