% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModuleEigen.R
\name{shinyModuleEigenUI}
\alias{shinyModuleEigenUI}
\alias{shinyModuleEigen}
\title{Shiny Module UI for Dendro Plot}
\usage{
shinyModuleEigenUI(id)

shinyModuleEigen(input, output, session, module_par, main_par, traitModule)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{traitModule}{reactive object with list created by `listof_wgcnamodules`}
}
\value{
nothing returned

reactive object for `shinyModuleEigenUI`
}
\description{
Shiny Module UI for Dendro Plot

Shiny Module Server for Dendro Plots
}
