% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModuleNames.R
\name{shinyModuleNamesUI}
\alias{shinyModuleNamesUI}
\alias{shinyModuleNames}
\title{Shiny Module UI for Module Names}
\usage{
shinyModuleNamesUI(id)

shinyModuleNames(input, output, session, module_par, traitModule, modrole)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{module_par}{reactive arguments from `foundrServer` and `shinyModules`}

\item{traitModule}{reactive object with list created by `listof_wgcnamodules`}

\item{modrole}{reactive with name of module role}
}
\value{
nothing returned

reactive object for `shinyModuleNamesUI`
}
\description{
Shiny Module UI for Module Names

Shiny Module Server for Dendro Plots
}
