% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModules.R
\name{shinyModulesUI}
\alias{shinyModulesUI}
\alias{shinyModules}
\title{Shiny Module UI for Modules Plot}
\usage{
shinyModulesUI(id)

shinyModules(
  input,
  output,
  session,
  main_par,
  traitDataInput,
  traitSignalInput,
  traitStatsInput
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{traitDataInput, traitSignalInput, traitStatsInput}{reactive objects from `foundrServer`}
}
\value{
nothing returned

reactive object for `shinyModulesUI`
}
\description{
Shiny Module UI for Modules Plot

Shiny Module Server for Modules Plots
}
