% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimePanel.R
\name{shinyTimePanelInput}
\alias{shinyTimePanelInput}
\alias{shinyTimePanelUI}
\alias{shinyTimePanelOutput}
\alias{shinyTimePanel}
\title{Shiny Module Input for Time Panel}
\usage{
shinyTimePanelInput(id)

shinyTimePanelUI(id)

shinyTimePanelOutput(id)

shinyTimePanel(id, main_par, traitData, traitSignal, traitStats)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitData, traitSignal, traitStats}{static objects}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Input for Time Panel

Shiny Module UI for Time Panel

Shiny Module Output for Times Plot

Shiny Module Server for Times Plots
}
