% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimePlot.R
\name{shinyTimePlotInput}
\alias{shinyTimePlotInput}
\alias{shinyTimePlotOutput}
\alias{shinyTimePlot}
\title{Shiny Module Input for Time Plots}
\usage{
shinyTimePlotInput(id)

shinyTimePlotOutput(id)

shinyTimePlot(id, main_par, traitData, traitSignal, traitTimesData)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitData, traitSignal, traitTimesData}{static objects}

\item{responses}{possible types of responses}
}
\value{
nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Input for Time Plots

Shiny Module Output for Time Plots

Shiny Module Server for Time Plots
}
