% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimesPanel.R
\name{shinyTimesPanel}
\alias{shinyTimesPanel}
\title{Shiny Module Server for Times Plots}
\usage{
shinyTimesPanel(id, main_par, traitData, traitSignalInput, traitStatsInput)
}
\arguments{
\item{main_par}{reactive arguments}

\item{traitData}{static objects}

\item{traitSignalInput, traitStatsInput}{reactive objects}
}
\value{
reactive object for `shinyTimesUI`
}
\description{
Shiny Module Server for Times Plots
}
