% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimeTable.R
\name{shinyTimeTableInput}
\alias{shinyTimeTableInput}
\alias{shinyTimeTableUI}
\alias{shinyTimeTableOutput}
\alias{shinyTimeTable}
\title{Shiny Module Input for Times Table}
\usage{
shinyTimeTableInput(id)

shinyTimeTableUI(id)

shinyTimeTableOutput(id)

shinyTimeTable(id, panel_par, main_par, traitData, traitSignal, traitStats)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitData, traitSignal, traitStats}{static objects}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Input for Times Table

Shiny Module UI for Times Table

Shiny Module Output for Times Table

Shiny Module Server for Times Plots
}
