% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimeTraits.R
\name{shinyTimeTraitsInput}
\alias{shinyTimeTraitsInput}
\alias{shinyTimeTraits}
\title{Shiny Module Input for Time Traits}
\usage{
shinyTimeTraitsInput(id)

shinyTimeTraits(
  id,
  panel_par,
  main_par,
  traitSignal,
  traitOrder,
  responses = c("value", "cellmean")
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitSignal}{static object}

\item{traitOrder}{reactive object}

\item{responses}{possible types of responses}
}
\value{
nothing returned

nothing returned
}
\description{
Shiny Module Input for Time Traits

Shiny Module Server for Times Plots
}
