% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitCors.R
\name{shinyTraitCorsUI}
\alias{shinyTraitCorsUI}
\alias{shinyTraitCors}
\title{Shiny Module UI for Trait Correlations}
\usage{
shinyTraitCorsUI(id)

shinyTraitCors(id, traitSignal, trait_names)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitSignal, traitStats}{reactive data frames}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

reactive object for `shinyTraitCorsUI`
}
\description{
Shiny Module UI for Trait Correlations

Shiny Module Server for Trait Stats
}
