% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitNames.R
\name{shinyTraitNamesUI}
\alias{shinyTraitNamesUI}
\alias{shinyTraitNames}
\title{Shiny Module UI for Trait Names}
\usage{
shinyTraitNamesUI(id)

shinyTraitNames(id, traitArranged, multiples = FALSE)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitArranged}{reactive data frames}

\item{multiples}{fixed logical for multiple trait names}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

reactive vector of trait names
}
\description{
Select trait names in one of two modes, depending on the fixed `multiples`:
`FALSE` = only one trait name,
`TRUE` =  multiple names.
The order of choices depends on `traitArranged()`.
}
