% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitObject.R
\name{shinyTraitObjectUI}
\alias{shinyTraitObjectUI}
\alias{shinyTraitObjectOutput}
\alias{shinyTraitObject}
\title{Shiny Module UI for Trait Object}
\usage{
shinyTraitObjectUI(id)

shinyTraitObjectOutput(id)

shinyTraitObject(
  input,
  output,
  session,
  main_par,
  trait_names,
  traitData,
  traitSignal
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{trait_names}{reactive with trait names.}

\item{traitData, traitSignal}{reactive objects from `foundrServer`}
}
\value{
nothing returned

nothing returned

reactive object for `shinyTrait` routines
}
\description{
Shiny Module UI for Trait Object

Shiny Module UI for Trait Object

Shiny Module Server for Trait Object
}
