% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitPairs.R
\name{shinyTraitPairsUI}
\alias{shinyTraitPairsUI}
\alias{shinyTraitPairs}
\title{Shiny Module UI for TraitPairs}
\usage{
shinyTraitPairsUI(id)

shinyTraitPairs(id, main_par, trait_names, traitSolosObject)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments from `foundrServer`}

\item{trait_names}{reactive with trait names}

\item{traitSolosObject}{reactive objects from `foundrServer`}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

reactive object for `shinyTaitSolosUI`
}
\description{
Shiny Module UI for TraitPairs

Shiny Module Server for trait solos Plots
}
