% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitPanel.R
\name{shinyTraitPanelInput}
\alias{shinyTraitPanelInput}
\alias{shinyTraitPanelOutput}
\alias{shinyTraitPanel}
\title{Shiny Module Input for Trait Panel}
\usage{
shinyTraitPanelInput(id)

shinyTraitPanelOutput(id)

shinyTraitPanel(
  id,
  main_par,
  traitData,
  traitSignal,
  traitStats,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitData, traitSignal, traitStats}{static data frames}

\item{customSettings}{list of custom settings}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Trait Panel

Shiny Module Output for Trait Panel

Shiny Module Server for Trait Panel
}
