% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitSolos.R
\name{shinyTraitSolo}
\alias{shinyTraitSolo}
\title{Shiny Module Server for trait solos Plots}
\usage{
shinyTraitSolo(
  input,
  output,
  session,
  main_par,
  datasets_selected,
  traitSignalInput
)
}
\arguments{
\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{datasets_selected}{traitSignalInput reactive objects from `foundrServer`}
}
\value{
reactive object for `shinyTaitSolosUI`
}
\description{
Shiny Module Server for trait solos Plots
}
