% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitSolos.R
\name{shinyTraitSolosUI}
\alias{shinyTraitSolosUI}
\alias{shinyTraitSolos}
\title{Shiny Module UI for traitSolos}
\usage{
shinyTraitSolosUI(id)

shinyTraitSolos(input, output, session, main_par, traitSolosObject)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{traitSolosObject}{reactive objects from `foundrServer`}
}
\value{
nothing returned

reactive object
}
\description{
Shiny Module UI for traitSolos

Shiny Module Server for trait solos Plots
}
