% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitTable.R
\name{shinyTraitTableUI}
\alias{shinyTraitTableUI}
\alias{shinyTraitTableOutput}
\alias{shinyTraitTable}
\title{Shiny Module UI for Trait Table}
\usage{
shinyTraitTableUI(id)

shinyTraitTableOutput(id)

shinyTraitTable(
  id,
  panel_par,
  main_par,
  keyTrait,
  relTraits,
  traitData,
  traitSignal,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{keyTrait, relTraits}{reactives with trait names}

\item{traitData, traitSignal}{static objects}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

reactive object for `shinyTrait` routines
}
\description{
Shiny Module UI for Trait Table

Shiny Module UI for Trait Object

Shiny Module Server for Trait Object
}
