% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVolcano.R
\name{shinyVolcanoOutput}
\alias{shinyVolcanoOutput}
\alias{shinyVolcano}
\title{Shiny Module Output for Volcano Plot}
\usage{
shinyVolcanoOutput(id)

shinyVolcano(id, main_par, traitStats, customSettings = NULL, facet = FALSE)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments from `server()`}

\item{traitStats}{static data frame}

\item{customSettings}{list of custom settings}

\item{facet}{facet on `strain` if `TRUE`}
}
\value{
nothing returned

reactive object for `shinyVolcanoUI`
}
\description{
Shiny Module Output for Volcano Plot

Shiny Module Server for Volcano Plots
}
