% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_time.R
\name{stats_time}
\alias{stats_time}
\title{Stats over Time}
\usage{
stats_time(
  traitStats,
  traitnames = timetraits(traitStats, timecol)[1],
  response = c("p.value", "SD"),
  timecol = c("week", "minute", "minsum"),
  ...
)
}
\arguments{
\item{traitStats}{data frame with trait stats from `strainstats`}

\item{traitnames}{names of `dataset: trait`}

\item{response}{character string for type of response}

\item{timecol}{column to use for time}

\item{...}{additional parameters ignored}
}
\description{
Stats over Time
}
