% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainstats.R
\name{strainstats}
\alias{strainstats}
\alias{summary_strainstats}
\alias{summary.strainstats}
\title{Use Broom to Find Stats for Model Summaries}
\usage{
strainstats(
  object,
  signal = ifelse(is_condition, paste("strain * sex *", condition_name), "strain * sex"),
  rest = ifelse(is_condition, paste("strain * sex + sex *", condition_name), "sex"),
  calc_sd = TRUE,
  condition_name = "condition",
  ...
)

summary_strainstats(
  object,
  terms = NULL,
  stats = c("deviation", "log10.p"),
  model = c("parts", "terms"),
  threshold = c(SD = 1, p = 0.01),
  ...
)

\method{summary}{strainstats}(object, ...)
}
\arguments{
\item{object}{object of class `strainstats`}

\item{signal}{signal factor combination as string for `formula`}

\item{rest}{rest factor combination as string for `formula`}

\item{calc_sd}{calculate SDs by `term` if `TRUE` (default)}

\item{condition_name}{name of `condition` column if present.}

\item{...}{not used}

\item{terms}{terms to include (overrides `model` and `stats`)}

\item{stats}{choice of `deviation` or `log10.p`}

\item{model}{choice of model `parts` or `terms`}

\item{threshold}{named vector for `SD` and `p.value`}
}
\value{
data frame with summaries by trait

data frame
}
\description{
Use Broom to Find Stats for Model Summaries

Summary of Strain Statistics
}
\examples{
out <- strainstats(sampleData)
summary(out, "deviation", "parts")
summary(out, "log10.p", "terms")
}
