% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_cor.R
\name{subset_eigen_cor}
\alias{subset_eigen_cor}
\title{Subset of eigen_cor object}
\usage{
subset_eigen_cor(x, colorname, facetname, ...)
}
\arguments{
\item{x}{of class `eigen_cor`}

\item{colorname}{color name}

\item{facetname}{facet name}
}
\value{
data frame with colorname, facetname, correlation
}
\description{
Subset of eigen_cor object
}
