% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_trait_names.R
\name{subset_trait_names}
\alias{subset_trait_names}
\title{Subset Trait Data on Names}
\usage{
subset_trait_names(
  object,
  traitnames = NULL,
  remove_columns = FALSE,
  drop_united = TRUE,
  sep = ": "
)
}
\arguments{
\item{object}{data frame}

\item{traitnames}{character string}

\item{remove_columns}{remove columns `dataset` and `trait` if `TRUE`}

\item{drop_united}{drop `datatraits` if `TRUE`}

\item{sep}{separator string}
}
\value{
data frame
}
\description{
Trait names of form `dataset: trait` will be selected.
}
