% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_wgcnaModules.R
\name{summary.listof_wgcnaModules}
\alias{summary.listof_wgcnaModules}
\alias{summary_wgcnaModules}
\alias{summary.wgcnaModules}
\title{Summary of List of WGCNA Modules}
\usage{
\method{summary}{listof_wgcnaModules}(object, ...)

summary_wgcnaModules(object, ...)

\method{summary}{wgcnaModules}(object, ...)
}
\arguments{
\item{object}{object of class `listof_wgcnaModules`}

\item{...}{additional parameters}

\item{object, x}{object of class `wgcnaModules`}

\item{main}{title for plot}
}
\value{
data frame
}
\description{
Summary of List of WGCNA Modules
}
