% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetraits.R
\name{timetraits}
\alias{timetraits}
\title{Get time trait roots}
\usage{
timetraits(
  object,
  timecol = c("week", "minute", "week_summary", "minute_summary")
)
}
\arguments{
\item{object}{data frame}

\item{timecol}{time column}
}
\description{
Get time trait roots
}
