% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitAncova.R
\name{traitAncova}
\alias{traitAncova}
\alias{summary_traitAncova}
\alias{ggplot_traitAncova}
\title{Trait Analysis of Covariance}
\usage{
traitAncova(
  traitData,
  traitStats,
  covarData,
  focus = c("Driver", "Target"),
  trait_name,
  condition_name = "diet"
)

summary_traitAncova(traitAncova, traits = NULL)

ggplot_traitAncova(
  traitAncova,
  traits = NULL,
  signif_level = 0.05,
  width = 0.1,
  height = 0
)
}
\arguments{
\item{traitData}{data frame of trait data}

\item{traitStats}{data frame of trait stats}

\item{covarData}{data frame of covariate data}

\item{focus}{focus of mediation}

\item{trait_name}{character string of focal trait name}

\item{condition_name}{character string of condition name}

\item{traitAncova}{object of class `traitAncova`}

\item{traits}{trait names to give unique colors}

\item{signif_level}{significance level}

\item{width, height}{jitter offsets}
}
\value{
object of class `traitAncova`

data frame as summary

ggplot object
}
\description{
Trait Analysis of Covariance

Summary of Trait Ancova

GGplot of Trait Ancova
}
