% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainstats.R
\name{traitOrderStats}
\alias{traitOrderStats}
\title{Order Stats by Selected term name}
\usage{
traitOrderStats(object, termname)
}
\arguments{
\item{object}{data frame from `strainstats`}

\item{termname}{name of element in `term` column to order by}
}
\description{
Order Stats by Selected term name
}
