% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitSolos.R
\name{traitSolos}
\alias{traitSolos}
\alias{summary_traitSolos}
\alias{summary.traitSolos}
\title{Select Traits and Strains for Response}
\usage{
traitSolos(
  traitData,
  traitSignal,
  traitnames = unique(unite_datatraits(traitSignal)),
  response = c("individual", "cellmean", "signal", "ind_signal"),
  strains = names(foundr::CCcolors),
  abbrev = FALSE,
  sep = ": "
)

summary_traitSolos(object, customSettings = NULL, ...)

\method{summary}{traitSolos}(object, ...)
}
\arguments{
\item{traitData}{data frame}

\item{traitSignal}{data frame}

\item{traitnames}{names of `dataset: trait` combinations to subset}

\item{response}{name of response to return}

\item{strains}{names of strains to subset}

\item{abbrev}{abbreviate names if `TRUE`}

\item{object}{object of class `traitSolos`}

\item{customSettings}{list of custom settings (including "condition_name" and "dataset")}

\item{...}{additional parameters}
}
\value{
data frame

summary table

summary table
}
\description{
Select Traits and Strains for Response

Summary of traitSolos object

Summary of traitSolos object
}
