% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_datatraits.R
\name{unite_datatraits}
\alias{unite_datatraits}
\title{Unite Dataset and Trait}
\usage{
unite_datatraits(
  object,
  traitnames,
  undo = FALSE,
  sep = ": ",
  filters = NULL,
  key = FALSE
)
}
\arguments{
\item{object}{data frame}

\item{traitnames}{names as `dataset: trait` for subsetting (when `undo` = `TRUE`)}

\item{undo}{logical flag on function use}

\item{sep}{separator for `dataset: trait` (default ": ")}

\item{filters}{optional list of columns to filter on (when `undo` = `FALSE`)}

\item{key}{get key traits if `TRUE`}
}
\value{
either a subset `object` based on `traitnames` = `dataset: trait` (if `undo` = `TRUE`)
        or vector of `dataset: trait` names (if `undo` = `FALSE`, default)
}
\description{
This operates in different ways based on `undo`.
Need to use both `dataset` and `trait` to identify data for a `trait`.
}
