% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{variation_biplot}
\alias{variation_biplot}
\title{BiPlot for Sources of Variation}
\usage{
variation_biplot(
  object,
  ordername,
  xlab,
  threshold,
  strain = "NONE",
  axes = 1:2,
  terms = termStats(object, signal = FALSE, drop_noise = FALSE),
  ...
)
}
\arguments{
\item{object}{data frame}

\item{ordername}{name of order column}

\item{xlab}{x label}

\item{threshold}{vector of threshold values}

\item{strain}{strain name to highlight in biplot}

\item{axes}{axes to plot}

\item{...}{additional paramaters ignored}
}
\value{
gg plot object
}
\description{
BiPlot for Sources of Variation
}
