% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{One-sided Volcano Plot}
\usage{
volcano(
  object,
  term = p_terms,
  threshold = c(sd = 0.6, p = 0.05),
  interact = FALSE,
  rescaleSD = TRUE
)
}
\arguments{
\item{object}{data frame with trait names, p-value and sd terms}

\item{term}{name of term to show}

\item{threshold}{named vector for `sd` and `p`}

\item{interact}{prepare for interactive if `TRUE`}

\item{rescaleSD}{rescale SD by `rawSD` if `TRUE`}
}
\description{
One-sided Volcano Plot
}
