% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abscorcalc.R
\name{abscorcalc}
\alias{abscorcalc}
\title{Calculate Absolute Correlations}
\usage{
abscorcalc(
  object,
  design = c("strain", "number", "sex", "diet"),
  method = "spearman"
)
}
\arguments{
\item{design}{names of design columns}

\item{method}{method for correlation}
}
\value{
data frame with design columns, `trait` names and `value` of absolute correlations
}
\description{
Calculate Absolute Correlations
}
