% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_under_curve.R
\name{area_under_curve}
\alias{area_under_curve}
\title{Calculate area under curve over minutes}
\usage{
area_under_curve(object, sum_function = auc)
}
\arguments{
\item{object}{data frame with `value` and other columns}
}
\value{
object collapsed over minutes with AUC and other summaries
}
\description{
Calculate area under curve over minutes
}
