% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestcor.R
\name{bestcor}
\alias{bestcor}
\title{Traits with Best Correlation}
\usage{
bestcor(object, traits, term = c("signal", "mean"))
}
\arguments{
\item{object}{data frame from `partition`}

\item{traits}{names of traits in `object`}

\item{term}{either `signal` or `mean`}
}
\value{
sorted vector of absolute correlations with names
}
\description{
Find other traits with best correlation with selected traits.
}
