% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestcor.R
\name{bestcor}
\alias{bestcor}
\alias{is_bestcor}
\alias{ggplot_bestcor}
\alias{autoplot.bestcor}
\alias{summary_bestcor}
\alias{summary.bestcor}
\title{Traits with Best Correlation}
\usage{
bestcor(traitSignal, traitnames = NULL, term = c("cellmean", "signal"))

is_bestcor(object)

ggplot_bestcor(object, mincor = 0.7, abscor = TRUE, ...)

\method{autoplot}{bestcor}(object, ...)

summary_bestcor(object, mincor = 0.5, ...)

\method{summary}{bestcor}(object, ...)
}
\arguments{
\item{traitSignal}{data frame from `partition`}

\item{traitnames}{names of traits in `traitSignal`}

\item{term}{either `signal` or `mean`}

\item{object}{object of class `bestcor`}

\item{mincor}{minimum correlation to show}

\item{abscor}{plot absolute value of correlation if `TRUE`}

\item{...}{not used}
}
\value{
sorted vector of absolute correlations with names

logical

ggplot object

data frame
}
\description{
Find other traits with best Spearman correlation with selected traitnames.
}
\examples{
sampleSignal <- partition(sampleData)
out <- bestcor(sampleSignal, "sample: A")
ggplot_bestcor(out, 0)
ggplot_bestcor(out, 0, abscor = FALSE)
}
