% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biggplot}
\alias{biggplot}
\title{Biplot using ggplot2 via ordr package}
\usage{
biggplot(bip_pca, axes = 1:2, scale.factor = 2)
}
\arguments{
\item{bip_pca}{object of class bip_pca}

\item{axes}{axes to plot}

\item{scale.factor}{scale factor for arrows vs points}
}
\value{
object of class ggplot2
}
\description{
Biplot using ggplot2 via ordr package
}
