% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot_data}
\alias{biplot_data}
\title{Prepare biplot data}
\usage{
biplot_data(dat, traits, factors = c("strain", "number", "sex", "diet"))
}
\arguments{
\item{dat}{}

\item{traits}{}

\item{factors}{}
}
\value{
data frame
}
\description{
Prepare biplot data
}
