% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot_pca}
\alias{biplot_pca}
\title{Get PCA components from biplot data}
\usage{
biplot_pca(
  bip,
  factors = c("strain", "number", "sex", "diet"),
  strain = "strain"
)
}
\arguments{
\item{bip}{}

\item{factors}{}

\item{strain}{}
}
\value{
data frame of PCA components
}
\description{
Get PCA components from biplot data
}
