% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot.R
\name{biplot_pca}
\alias{biplot_pca}
\title{Get PCA components from biplot data}
\usage{
biplot_pca(bip, size = c("module", "kME", "p.value", "size"))
}
\arguments{
\item{bip}{object from `biplot_data`}

\item{transpose}{transpose for `princomp` if `TRUE`}
}
\value{
data frame of PCA components
}
\description{
Get PCA components from biplot data
}
