% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broomit.R
\name{broomit}
\alias{broomit}
\title{Use Broom to Find P-values for Model Summaries}
\usage{
broomit(
  object,
  trait = "trait",
  value = "value",
  signal = "strain * sex * diet",
  design = "strain * sex + sex * diet",
  interact = stringr::str_remove(signal, " .*")
)
}
\arguments{
\item{object}{data frame in long format with trait data}

\item{trait}{name of column with trait names}

\item{value}{name column with trait values}

\item{signal}{signal factor combination as string for `formula`}

\item{design}{design factor combination as string for `formula`}

\item{interact}{interacting factor to select for summary table}
}
\value{
data frame with summaries by trait
}
\description{
Use Broom to Find P-values for Model Summaries
}
