% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionContrasts.R
\name{conditionContrasts}
\alias{conditionContrasts}
\alias{ggplot_conditionContrasts}
\alias{plot.conditionContrasts}
\alias{summary_conditionContrasts}
\alias{summary.conditionContrasts}
\alias{c.conditionContrasts}
\alias{split.conditionContrasts}
\title{Contrasts of Conditions}
\usage{
conditionContrasts(
  traitSignal,
  traitStats,
  termname = "signal",
  rawStats = traitStats
)

ggplot_conditionContrasts(
  object,
  bysex = sexes,
  ntraits = 20,
  plottype = c("dotplot", "volcano", "biplot"),
  ordername = attr(object, "ordername"),
  ...
)

\method{plot}{conditionContrasts}(x, ...)

summary_conditionContrasts(
  object,
  ntrait = 20,
  sortby = ordername,
  ordername = attr(object, "ordername"),
  ...
)

\method{summary}{conditionContrasts}(object, ...)

\method{c}{conditionContrasts}(...)

\method{split}{conditionContrasts}(x, ...)
}
\arguments{
\item{traitSignal}{data frame of signals}

\item{traitStats}{data frame of stats}

\item{termname}{name of term}

\item{rawStats}{data frame with `rawSD`}

\item{object}{object of class `conditionContrasts`}

\item{bysex}{type of sex from c("F","M","F-M","F+M")}

\item{ntraits}{number of traits to plot}

\item{plottype}{one of "dotplot", "volcano" or "biplot"}

\item{ordername}{column name to order entries}

\item{...}{additional parameters for generic `split`}

\item{x}{objects of class `conditionContrasts`}

\item{ntrait}{number of traits to plot}

\item{sortby}{sort summary by this variable}
}
\value{
object of class `conditionContrasts`

ggplot object

ggplot object

data frame

data frame

object of class `conditionContrasts`

object of class `conditionContrasts`
}
\description{
Contrasts of Conditions

GGplot of Contrasts of Conditions

Plot method for Contrasts of Conditions

Summary method for Contrasts of Condtions

Summary method for Contrasts of Conditions

Combine method for Contrasts of Conditions

Split method for Contrasts of Conditions
}
