% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditionContrasts.R
\name{conditionContrasts}
\alias{conditionContrasts}
\alias{ggplot_conditionContrasts}
\alias{plot.conditionContrasts}
\alias{summary_conditionContrasts}
\alias{summary.conditionContrasts}
\title{Contrasts of Conditions}
\usage{
conditionContrasts(
  traitSignal,
  traitStats,
  termname = "signal",
  rawStats = traitStats
)

ggplot_conditionContrasts(
  object,
  bysex = names(sexes),
  ntraits = 20,
  volcano = FALSE,
  ...
)

\method{plot}{conditionContrasts}(x, ...)

summary_conditionContrasts(object, ntrait = 20, ...)

\method{summary}{conditionContrasts}(object, ...)
}
\arguments{
\item{traitSignal}{data frame of signals}

\item{traitStats}{data frame of stats}

\item{termname}{name of term}

\item{rawStats}{data frame with `rawSD`}

\item{object}{object of class `conditionContrasts`}

\item{bysex}{type of sex from c("F","M","F-M","F+M")}

\item{ntraits}{number of traits to plot}

\item{volcano}{volcano plot if `TRUE`}

\item{...}{parameters passed to `ggplot_conditionContrasts`}

\item{x}{object of class `conditionContrasts`}

\item{ntrait}{number of traits to plot}
}
\value{
object of class `conditionContrasts`

ggplot object

ggplot object

data frame

data frame
}
\description{
Contrasts of Conditions

GGplot of Contrasts of Conditions

Plot method for Contrasts of Condtions

Summary method for Contrasts of Condtions

Summary method for Contrasts of Condtions
}
