% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectplot.R
\name{effectplot}
\alias{effectplot}
\title{Effect Plots}
\usage{
effectplot(object, traitnames = unique(object$trait)[1:5])
}
\arguments{
\item{object}{data frame with `trait` names, `term` names, `SD` and `p.value`}

\item{traitnames}{`traitnames` to show}
}
\description{
Effect Plots
}
