% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectplot.R
\name{effectplot}
\alias{effectplot}
\title{Effect Plots}
\usage{
effectplot(object, traitnames = NULL, correlated = NULL, ...)
}
\arguments{
\item{object}{data frame from `strainstats()`}

\item{traitnames}{`traitnames` to show}

\item{correlated}{correlated `traitnames` to overlay on boxplot}

\item{...}{additional parameters}
}
\description{
Effect Plots
}
\examples{
sampleStats <- strainstats(sampleData)
effectplot(sampleStats, trait_names(sampleStats, "C"))

}
