% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{foundrServer}
\alias{foundrServer}
\title{Server for Founder Shiny App}
\usage{
foundrServer(input, output, session, traitdat, traitsumdat, condition)
}
\arguments{
\item{input, output, session}{shiny parameters}

\item{traitdat}{data frame with trait data}

\item{traitsumdat}{data frame with summary data}

\item{condition}{column(s) identifying condition for plotting}
}
\value{
A Server definition that can be passed to the `shinyServer` function.
}
\description{
User needs to supply the following routines:
  foundrIntro() # introductory remarks about study
  foundrSettings() # settings to be used in later parts
  foundrScatplot() # wrapper for foundr::scatplot
  foundrData() # routine to subset data
  foundrMean() # routine to subset mean summaries
}
