% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{foundrServer}
\alias{foundrServer}
\title{Server for Founder Shiny App}
\usage{
foundrServer(
  input,
  output,
  session,
  traitdata = NULL,
  traitstats = NULL,
  traitsignal = NULL,
  customSettings = NULL,
  condition_name = "condition"
)
}
\arguments{
\item{input, output, session}{shiny parameters}

\item{traitdata}{data frame with trait data}

\item{traitstats}{data frame with summary data}

\item{customSettings}{list of custom settings}

\item{condition_name}{name of condition}
}
\value{
A Server definition that can be passed to the `shinyServer` function.
}
\description{
The `customSettings` is a list that may include
  `help` = name of help markdown file (e.g. `"help.md"`)
  `condition` = names of condition (e.g. `"diet"`) 
  `datatype` = names of datatypes (e.g. `uploaded = "MyUpload"`)
}
\details{
User needs to supply the following routines:
  userDatasets() # function returning list of text for intro
}
