% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{foundrServer}
\alias{foundrServer}
\title{Server for Founder Shiny App}
\usage{
foundrServer(
  input,
  output,
  session,
  traitdata = NULL,
  traitstats = NULL,
  traitsignal = NULL,
  helppath = "",
  condition_name = "condition"
)
}
\arguments{
\item{input, output, session}{shiny parameters}

\item{traitdata}{data frame with trait data}

\item{traitstats}{data frame with summary data}

\item{condition_name}{name of condition}

\item{helpath}{path to markdown file (say `help.md`) for intro}
}
\value{
A Server definition that can be passed to the `shinyServer` function.
}
\description{
User needs to supply the following routines:
  userDatasets() # function returning list of text for intro
}
