% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_time.R
\name{ggplot_strain_time}
\alias{ggplot_strain_time}
\title{GGplot of Strains over Time}
\usage{
ggplot_strain_time(object, ...)
}
\arguments{
\item{object}{object of class `strain_time`}

\item{...}{additional parameters}
}
\value{
ggplot object
}
\description{
GGplot of Strains over Time
}
