% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_strain_time.R
\name{ggplot_strain_time}
\alias{ggplot_strain_time}
\title{GGplot of Strains over Time}
\usage{
ggplot_strain_time(
  traitData,
  traitSignal,
  datatraits,
  timecol = "week",
  response = c("value", "cellmean", "signal"),
  ...
)
}
\arguments{
\item{traitData}{data frame}

\item{traitSignal}{data frame of summary signals}

\item{datatraits}{`dataset: trait` name(s) (typically one)}

\item{timecol}{column to use for time}

\item{response}{type of response}

\item{...}{additional parameters}
}
\value{
ggplot object
}
\description{
GGplot of Strains over Time
}
