% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_template.R
\name{ggplot_template}
\alias{ggplot_template}
\title{GGplot template for other routines}
\usage{
ggplot_template(object, ..., drop_xlab = FALSE, legend_position = "none")
}
\arguments{
\item{object, x}{data frame to be plotted}

\item{...}{additional parameters}

\item{drop_xlab}{drop `xlab` except last if `TRUE`}

\item{legend_position}{show legend if `TRUE`}

\item{facet_strain}{facet by strain if `TRUE`}

\item{shape_sex}{use different shape by sex if `TRUE`}

\item{boxplot}{overlay boxplot if `TRUE`}

\item{horizontal}{flip vertical and horizontal axis if `TRUE`}
}
\value{
object of class ggplot
}
\description{
GGplot template for other routines
}
