% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_time.R
\name{ggplot_time}
\alias{ggplot_time}
\title{GGplot of object over time}
\usage{
ggplot_time(object)
}
\arguments{
\item{object}{}
}
\value{
object of class ggplot2
}
\description{
GGplot of object over time
}
