% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitPairs.R
\name{ggplot_traitPairs}
\alias{ggplot_traitPairs}
\title{GGplot of trait pairs}
\usage{
ggplot_traitPairs(object, ...)
}
\arguments{
\item{object, x}{object of class `traitPairs`}

\item{...}{}
}
\value{
ggplot object
}
\description{
GGplot of trait pairs
}
