% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleHarmony.R
\name{moduleHarmony}
\alias{moduleHarmony}
\title{Harmonize module object}
\usage{
moduleHarmony(
  datasetname,
  links,
  object,
  responsename = names(object[[1]]),
  ...
)
}
\arguments{
\item{datasetname}{name to give to dataset (here for compatibility)}

\item{links}{data frame with links (not used yet)}

\item{object}{object of class traitModule}

\item{responsename}{name of response to extract}

\item{...}{additional parameters ignored}
}
\value{
harmonize object
}
\description{
Harmonize module object
}
