% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleHarmony.R
\name{moduleHarmony}
\alias{moduleHarmony}
\title{Harmonize module object}
\usage{
moduleHarmony(datasetname, links, object, response = names(object), ...)
}
\arguments{
\item{links}{data frame with links (not used yet)}

\item{object}{object of class traitModule}

\item{response}{name of response to extract}

\item{...}{additional parameters ignored}
}
\value{
harmonize object
}
\description{
Harmonize module object
}
