% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition Trait into Signal, Ancillary and Noise}
\usage{
partition(
  object,
  trait = "trait",
  value = "value",
  signal = "strain * sex * condition",
  ancillary = "strain * sex + sex * condition"
)
}
\arguments{
\item{object}{data frame in long format with trait data}

\item{trait}{name of column with trait names}

\item{value}{name column with trait values}

\item{signal}{signal factor combination as string for `formula`}

\item{ancillary}{ancillary factor combination as string for `formula`}
}
\value{
data frame with added columns `ancillary`, `signal`, `noise`
}
\description{
Partition a trait value into three parts based on ancillary and signal factors.
The signal is the main interest for comparison across traits, while the ancillary
concerns aspects of the experiment that are controlled. Noise is unexplained variation.
The three parts will be orthogonal to each other.
The `signal` factors, entered as terms for `formula`, are conditioned by the `ancillary` factors.
}
