% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatplot.R
\name{scatplot}
\alias{scatplot}
\title{Scatter plot by trait values}
\usage{
scatplot(
  data,
  x,
  y,
  shape_sex = TRUE,
  line_strain = TRUE,
  title = paste(x, "vs", y)
)
}
\arguments{
\item{data}{data frame}

\item{x}{name of first trait}

\item{y}{name of second trait}

\item{shape_sex}{use different shape by `sex` if `TRUE`}

\item{line_strain}{show separate lines by `strain` if `TRUE`}

\item{title}{title for plot}
}
\description{
Scatter plot by trait values
}
\examples{
scatplot(sampleData, "A", "B")
}
