% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_time.R
\name{separate_time}
\alias{separate_time}
\title{Separate Time Traits}
\usage{
separate_time(
  object,
  traitnames,
  timeunit = c("week", "minute", "minute_summary", "week_summary")
)
}
\arguments{
\item{object}{data frame}

\item{traitnames}{character vector with `dataset: trait` names}

\item{timeunit}{name of time unit}
}
\value{
data frame
}
\description{
Separate Time Traits
}
