% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny Server for foundr Package}
\usage{
server(
  input,
  output,
  session,
  traitData = NULL,
  traitSignal = NULL,
  traitStats = NULL,
  customSettings = NULL
)
}
\arguments{
\item{input, output, session}{shiny server parameters}

\item{traitData, traitSignal, traitStats}{trait data frame}

\item{customSettings}{list of custom settings}
}
\value{
reactive server
}
\description{
Shiny Server for foundr Package
}
