% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastModule.R
\name{shinyContrastModule}
\alias{shinyContrastModule}
\title{Shiny Module Server for Modules of Contrasts}
\usage{
shinyContrastModule(
  id,
  panel_par,
  main_par,
  traitContrast,
  contrastModule,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitContrast, contrastModule}{static data frames}

\item{customSettings}{list of custom settings}
}
\value{
reactive object
}
\description{
Shiny Module Server for Modules of Contrasts
}
