% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastPanel.R
\name{shinyContrastPanelInput}
\alias{shinyContrastPanelInput}
\alias{shinyContrastPanelUI}
\alias{shinyContrastPanelOutput}
\alias{shinyContrastPanel}
\title{Shiny Module Input for Contrast Panel}
\usage{
shinyContrastPanelInput(id)

shinyContrastPanelUI(id)

shinyContrastPanelOutput(id)

shinyContrastPanel(
  id,
  main_par,
  traitSignal,
  traitStats,
  traitModule,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitSignal, traitStats, traitModule}{static objects}

\item{customSettings}{list of custom settings}
}
\value{
nothing returned

nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Contrast Panel

Shiny Module Input for Contrast Panel

Shiny Module Output for Contrast Panel

Shiny Module Server for Contrast Panel
}
