% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastPlot.R
\name{shinyContrastPlotInput}
\alias{shinyContrastPlotInput}
\alias{shinyContrastPlotOutput}
\alias{shinyContrastPlot}
\title{Shiny Module Input for Contrast Plots}
\usage{
shinyContrastPlotInput(id)

shinyContrastPlotOutput(id)

shinyContrastPlot(id, main_par, contrastTable, customSettings = NULL)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Contrast Plots

Shiny Module Output for Contrast Plots

Shiny Module Server for Contrast Plots
}
