% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastPlot.R
\name{shinyContrastPlotInput}
\alias{shinyContrastPlotInput}
\alias{shinyContrastPlotUI}
\alias{shinyContrastPlotOutput}
\alias{shinyContrastPlot}
\title{Shiny Module Input for Contrast Plots}
\usage{
shinyContrastPlotInput(id)

shinyContrastPlotUI(id)

shinyContrastPlotOutput(id)

shinyContrastPlot(
  id,
  sex_par,
  panel_par,
  main_par,
  contrastTable,
  customSettings = NULL,
  modTitle = shiny::reactive("Eigentrait Contrasts")
)
}
\arguments{
\item{id}{identifier}

\item{sex_par, panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}

\item{modTitle}{character string title for section}
}
\value{
nothing returned

nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Contrast Plots

Shiny Module UI for Contrast Plots

Shiny Module Output for Contrast Plots

Shiny Module Server for Contrast Plots
}
