% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastSex.R
\name{shinyContrastSexInput}
\alias{shinyContrastSexInput}
\alias{shinyContrastSexUI}
\alias{shinyContrastSexOutput}
\alias{shinyContrastSex}
\title{Shiny Sex Input for Contrast Plots}
\usage{
shinyContrastSexInput(id)

shinyContrastSexUI(id)

shinyContrastSexOutput(id)

shinyContrastSex(id, panel_par, main_par, contrastTable, customSettings = NULL)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{input parameters}

\item{contrastTable}{reactive data frame}

\item{customSettings}{list of custom settings}
}
\value{
nothing returned

nothing returned

nothing returned

reactive object
}
\description{
Shiny Sex Input for Contrast Plots

Shiny Sex UI for Contrast Plots

Shiny Sex Output for Contrast Plots

Shiny Sex Server for Contrast Plots
}
