% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastTime.R
\name{shinyContrastTimeInput}
\alias{shinyContrastTimeInput}
\alias{shinyContrastTimeUI}
\alias{shinyContrastTime}
\title{Shiny Module Input for Contrasts over Time}
\usage{
shinyContrastTimeInput(id)

shinyContrastTimeUI(id)

shinyContrastTime(
  id,
  panel_par,
  main_par,
  traitSignal,
  traitStats,
  contrastTable,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitSignal, traitStats}{static data frames}

\item{customSettings}{list of custom settings}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Contrasts over Time

Shiny Module UI for Contrasts over Time

Shiny Module Server for Contrasts over Time
}
