% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastTime.R
\name{shinyContrastTimeInput}
\alias{shinyContrastTimeInput}
\alias{shinyContrastTimeOutput}
\alias{shinyContrastTime}
\title{Shiny Module Input for Contrasts over Time}
\usage{
shinyContrastTimeInput(id)

shinyContrastTimeOutput(id)

shinyContrastTime(id, main_par, traitSignal, traitStats, customSettings = NULL)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitSignal, traitStats}{static data frames}

\item{customSettings}{list of custom settings}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Contrasts over Time

Shiny Module Output for Contrasts over Time

Shiny Module Server for Contrasts over Time
}
