% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCorPlot.R
\name{shinyCorPlotUI}
\alias{shinyCorPlotUI}
\alias{shinyCorPlotOutput}
\alias{shinyCorPlot}
\title{Shiny Module UI for Trait Correlations}
\usage{
shinyCorPlotUI(id)

shinyCorPlotOutput(id)

shinyCorPlot(id, panel_par, main_par, CorTable)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive inputs from calling modules}

\item{CorTable}{reactive data frames}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module UI for Trait Correlations

Shiny Module UI for Trait Correlations

Shiny Module Server for Trait Stats
}
