% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCorTable.R
\name{shinyCorTableInput}
\alias{shinyCorTableInput}
\alias{shinyCorTableUI}
\alias{shinyCorTableOutput}
\alias{shinyCorTable}
\title{Shiny Module Input for Trait Correlations}
\usage{
shinyCorTableInput(id)

shinyCorTableUI(id)

shinyCorTableOutput(id)

shinyCorTable(id, main_par, traits_par, traitArranged, traitSignal)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par, traits_par}{reactive inputs from calling modules}

\item{traitArranged, traitSignal}{reactive data frames}
}
\value{
nothing returned

nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Trait Correlations

Shiny Module UI for Trait Correlations

Shiny Module UI for Trait Correlations

Shiny Module Server for Trait Stats
}
