% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyCorTable.R
\name{shinyCorTableUI}
\alias{shinyCorTableUI}
\alias{shinyCorTableOutput}
\alias{shinyCorTable}
\title{Shiny Module UI for Trait Correlations}
\usage{
shinyCorTableUI(id)

shinyCorTableOutput(id)

shinyCorTable(
  id,
  main_par,
  traits_par,
  keyTrait,
  traitSignal,
  customSettings = NULL
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par, traits_par}{reactive inputs from calling modules}

\item{keyTrait}{reactive character string}

\item{traitSignal}{static data frame}

\item{customSettings}{list of custom settings}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module UI for Trait Correlations

Shiny Module UI for Trait Correlations

Shiny Module Server for Trait Stats
}
