% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimePlot.R
\name{shinyTimePlotOutput}
\alias{shinyTimePlotOutput}
\alias{shinyTimePlot}
\title{Shiny Module Output for Time Plots}
\usage{
shinyTimePlotOutput(id)

shinyTimePlot(id, panel_par, main_par, traitSignal, traitTimesData)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitSignal}{static object}

\item{traitTimesData}{reactive object}

\item{responses}{possible types of responses}
}
\value{
nothing returned

nothing returned
}
\description{
Shiny Module Output for Time Plots

Shiny Module Server for Time Plots
}
