% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimes.R, R/shinyTimesPanel.R
\name{shinyTimesUI}
\alias{shinyTimesUI}
\alias{shinyTimes}
\alias{shinyTimesPanelUI}
\title{Shiny Module UI for Times Plot}
\usage{
shinyTimesUI(id)

shinyTimes(
  input,
  output,
  session,
  main_par,
  traitDataInput,
  traitSignalInput,
  traitStatsInput
)

shinyTimesPanelUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{traitDataInput, traitSignalInput, traitStatsInput}{reactive objects from `foundrServer`}
}
\value{
nothing returned

reactive object for `shinyTimesUI`

nothing returned
}
\description{
Shiny Module UI for Times Plot

Shiny Module Server for Times Plots

Shiny Module UI for Times Plot
}
