% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimesPanel.R
\name{shinyTimesPanelInput}
\alias{shinyTimesPanelInput}
\alias{shinyTimesPanelUI}
\alias{shinyTimesPanelOutput}
\alias{shinyTimesPanel}
\title{Shiny Module Input for Times Plot}
\usage{
shinyTimesPanelInput(id)

shinyTimesPanelUI(id)

shinyTimesPanelOutput(id)

shinyTimesPanel(id, main_par, traitData, traitSignalInput, traitStatsInput)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitData}{static objects}

\item{traitSignalInput, traitStatsInput}{reactive objects}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Input for Times Plot

Shiny Module UI for Times Plot

Shiny Module Output for Times Plot

Shiny Module Server for Times Plots
}
