% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTimeTable.R
\name{shinyTimeTableInput}
\alias{shinyTimeTableInput}
\alias{shinyTimeTableOutput}
\alias{shinyTimeTable}
\title{Shiny Module Input for Times Table}
\usage{
shinyTimeTableInput(id)

shinyTimeTableOutput(id)

shinyTimeTable(
  id,
  main_par,
  traitData,
  traitSignal,
  traitStats,
  responses = c("value", "cellmean")
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{traitData, traitSignal, traitStats}{static objects}

\item{responses}{possible types of responses}
}
\value{
nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Input for Times Table

Shiny Module Output for Times Plot

Shiny Module Server for Times Plots
}
