% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyContrastModule.R, R/shinyContrastTime.R,
%   R/shinyTimeTraits.R
\name{shinyContrastModuleInput}
\alias{shinyContrastModuleInput}
\alias{shinyContrastModuleOutput}
\alias{shinyContrastTimeInput}
\alias{shinyTimeTraitsInput}
\alias{shinyTimeTraits}
\title{Shiny Module Input for Modules of Contrasts}
\usage{
shinyContrastModuleInput(id)

shinyContrastModuleOutput(id)

shinyContrastTimeInput(id)

shinyTimeTraitsInput(id)

shinyTimeTraits(
  id,
  panel_par,
  main_par,
  traitSignal,
  traitOrder,
  responses = c("value", "cellmean")
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{reactive arguments}

\item{traitSignal}{static object}

\item{traitOrder}{reactive object}

\item{responses}{possible types of responses}
}
\value{
nothing returned

nothing returned

nothing returned

nothing returned

nothing returned
}
\description{
Shiny Module Input for Modules of Contrasts

Shiny Module Output for Modules of Contrasts

Shiny Module Input for Contrasts over Time

Shiny Module Input for Time Traits

Shiny Module Server for Times Plots
}
