% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitCors.R
\name{shinyTraitCorsUI}
\alias{shinyTraitCorsUI}
\alias{shinyTraitCorsOutput}
\alias{shinyTraitCors}
\title{Shiny Module UI for Trait Correlations}
\usage{
shinyTraitCorsUI(id)

shinyTraitCorsOutput(id)

shinyTraitCors(id, stats_par, module_par, key_trait, traitSignal)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{stats_par, module_par}{reactive inputs from calling modules}

\item{traitSignal, traitStats}{reactive data frames}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

reactive object for `shinyTraitCorsUI`
}
\description{
Shiny Module UI for Trait Correlations

Shiny Module UI for Trait Correlations

Shiny Module Server for Trait Stats
}
