% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitNames.R
\name{shinyTraitNamesUI}
\alias{shinyTraitNamesUI}
\alias{shinyTraitNames}
\title{Shiny Module UI for Trait Names}
\usage{
shinyTraitNamesUI(id)

shinyTraitNames(
  input,
  output,
  session,
  traitStats,
  traitStatsArranged,
  multiples = shiny::reactive(FALSE)
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{traitStats, traitStatsArranged}{reactive object with list created by `listof_wgcnamodules`}

\item{multiples}{reactive logical for multiple trait names}
}
\value{
nothing returned

reactive vector of trait names
}
\description{
Select trait names in one of two modes, depending on the fixed `multiples()`:
`FALSE` = only one trait name,
`TRUE` =  multiple names.
The order of choices depends on `traitStatsArranged()`,
which is decided in the `shinyTraitStats` module.
}
