% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitOrder.R
\name{shinyTraitOrderInput}
\alias{shinyTraitOrderInput}
\alias{shinyTraitOrderUI}
\alias{shinyTraitOrderOutput}
\alias{shinyTraitOrder}
\title{Shiny Module Input for Trait Stats}
\usage{
shinyTraitOrderInput(id)

shinyTraitOrderUI(id)

shinyTraitOrderOutput(id)

shinyTraitOrder(
  id,
  main_par,
  traitStats,
  traitSignal = NULL,
  customSettings = NULL,
  allDatasets = FALSE
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{input reactive list}

\item{traitStats, traitSignal}{static data frame}

\item{customSettings}{custom settings list}

\item{allDatasets}{initially select all datasets if `TRUE`}
}
\value{
nothing returned

nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Trait Stats

Shiny Module UI for Trait Stats

Shiny Module Output for Trait Stats

Shiny Module Server for Trait Stats
}
