% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitOrder.R
\name{shinyTraitOrderInput}
\alias{shinyTraitOrderInput}
\alias{shinyTraitOrderUI}
\alias{shinyTraitOrderOutput}
\alias{shinyTraitOrder}
\title{Shiny Module Input for Trait Stats}
\usage{
shinyTraitOrderInput(id)

shinyTraitOrderUI(id)

shinyTraitOrderOutput(id)

shinyTraitOrder(id, traitStats, traitSignal = NULL)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitStats}{reactive data frames}
}
\value{
nothing returned

nothing returned

nothing returned

reactive object
}
\description{
Shiny Module Input for Trait Stats

Shiny Module UI for Trait Stats

Shiny Module Output for Trait Stats

Shiny Module Server for Trait Stats
}
