% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitOrder.R
\name{shinyTraitOrderInput}
\alias{shinyTraitOrderInput}
\alias{shinyTraitOrderUI}
\alias{shinyTraitOrder}
\title{Shiny Module Input for Trait Stats}
\usage{
shinyTraitOrderInput(id)

shinyTraitOrderUI(id)

shinyTraitOrder(
  id,
  panel_par,
  main_par,
  traitStats,
  customSettings = NULL,
  allDatasets = FALSE
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{panel_par, main_par}{input reactive list}

\item{traitStats}{static data frame}

\item{customSettings}{custom settings list}

\item{allDatasets}{initially select all datasets if `TRUE`}
}
\value{
reactive object
}
\description{
Shiny Module Input for Trait Stats

Shiny Module UI for Trait Stats

Shiny Module Server for Trait Stats
}
