% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitPairs.R
\name{shinyTraitPairsUI}
\alias{shinyTraitPairsUI}
\alias{shinyTraitPairs}
\title{Shiny Module UI for TraitPairs}
\usage{
shinyTraitPairsUI(id)

shinyTraitPairs(
  input,
  output,
  session,
  main_par,
  trait_names,
  traitSolosObject
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{trait_names}{reactive with trait names.}

\item{traitSolosObject}{reactive objects from `foundrServer`}
}
\value{
nothing returned

reactive object for `shinyTaitSolosUI`
}
\description{
Shiny Module UI for TraitPairs

Shiny Module Server for trait solos Plots
}
