% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitPanel.R
\name{shinyTraitPanelUI}
\alias{shinyTraitPanelUI}
\alias{shinyTraitPanelOutput}
\alias{shinyTraitPanel}
\title{Shiny Module UI for Trait Panel}
\usage{
shinyTraitPanelUI(id)

shinyTraitPanelOutput(id)

shinyTraitPanel(id, module_par, traitData, traitSignal, traitStats)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{traitData}{static data frame}

\item{traitSignal, traitStats}{reactive data frames}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

reactive object
}
\description{
Shiny Module UI for Trait Panel

Shiny Module Output for Trait Panel

Shiny Module Server for Trait Panel
}
