% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitSolos.R
\name{shinyTraitSoloUI}
\alias{shinyTraitSoloUI}
\title{Shiny Module UI for traitSolos}
\usage{
shinyTraitSoloUI(id)
}
\arguments{
\item{id}{identifier for shiny reactive}
}
\value{
nothing returned
}
\description{
Shiny Module UI for traitSolos
}
