% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitStats.R
\name{shinyTraitStatsInput}
\alias{shinyTraitStatsInput}
\alias{shinyTraitStatsUI}
\alias{shinyTraitStatsOutput}
\alias{shinyTraitStats}
\title{Shiny Module UI for Trait Stats}
\usage{
shinyTraitStatsInput(id)

shinyTraitStatsUI(id)

shinyTraitStatsOutput(id)

shinyTraitStats(id, module_par, traitSignal, traitStats)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{module_par}{reactive input from calling module}

\item{traitSignal, traitStats}{reactive data frames}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

nothing returned

reactive object for `shinyTraitStatsUI`
}
\description{
Shiny Module UI for Trait Stats

Shiny Module UI for Trait Stats

Shiny Module Output for Trait Stats

Shiny Module Server for Trait Stats
}
