% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitStats.R
\name{shinyTraitStatsUI}
\alias{shinyTraitStatsUI}
\alias{shinyTraitStats}
\title{Shiny Module UI for Trait Stats}
\usage{
shinyTraitStatsUI(id)

shinyTraitStats(input, output, session, traitSignal, traitStats)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{traitSignal, traitStats}{reactive data frames}
}
\value{
nothing returned

reactive object for `shinyTraitStatsUI`
}
\description{
Shiny Module UI for Trait Stats

Shiny Module Server for Dendro Plots
}
