% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyTraitTable.R
\name{shinyTraitTableUI}
\alias{shinyTraitTableUI}
\alias{shinyTraitTableOutput}
\alias{shinyTraitTable}
\title{Shiny Module UI for Trait Table}
\usage{
shinyTraitTableUI(id)

shinyTraitTableOutput(id)

shinyTraitTable(
  id,
  main_par,
  trait_names,
  traitData,
  traitSignal,
  customSettings
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{main_par}{reactive arguments}

\item{trait_names}{reactive with trait names}

\item{traitData, traitSignal}{reactive objects from `foundrServer`}

\item{input, output, session}{standard shiny arguments}
}
\value{
nothing returned

nothing returned

reactive object for `shinyTrait` routines
}
\description{
Shiny Module UI for Trait Table

Shiny Module UI for Trait Object

Shiny Module Server for Trait Object
}
