% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyVolcano.R
\name{shinyVolcanoUI}
\alias{shinyVolcanoUI}
\alias{shinyVolcano}
\title{Shiny Module UI for Volcano Plot}
\usage{
shinyVolcanoUI(id)

shinyVolcano(
  input,
  output,
  session,
  main_par,
  traitStatsSelectType,
  traitStatsArranged
)
}
\arguments{
\item{id}{identifier for shiny reactive}

\item{input, output, session}{standard shiny arguments}

\item{main_par}{reactive arguments from `foundrServer`}

\item{traitStatsSelectType, traitStatsArranged}{reactive objects from `foundrServer`}
}
\value{
nothing returned

reactive object for `shinyVolcanoUI`
}
\description{
Shiny Module UI for Volcano Plot

Shiny Module Server for Volcano Plots
}
