% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_time.R
\name{strain_time}
\alias{strain_time}
\alias{stats_time}
\title{Strains over Time}
\usage{
strain_time(
  traitData,
  traitSignal,
  traitnames = timetraits(traitSignal, timecol)[1],
  response = c("value", "cellmean", "signal"),
  timecol = c("week", "minute"),
  ...
)

stats_time(
  traitStats,
  traitnames = timetraits(traitStats, timecol)[1],
  response = c("p.value", "SD"),
  timecol = c("week", "minute"),
  ...
)
}
\arguments{
\item{traitData}{data frame with trait data}

\item{traitSignal}{data from with trait signals}

\item{traitnames}{names of `dataset: trait`}

\item{response}{character string for type of response}

\item{timecol}{column to use for time}

\item{...}{additional parameters ignored}

\item{traitStats}{data frame with trait stats}
}
\value{
object of class `strain_time`

object of class `strain_time`
}
\description{
Strains over Time

Stats over Time
}
