% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitTimes.R
\name{strain_time}
\alias{strain_time}
\title{Strains over Time}
\usage{
strain_time(
  traitData,
  traitSignal,
  traitnames = timetraits(traitSignal, timecol)[1],
  response = c("value", "cellmean", "signal"),
  timecol = c("week", "minute", "minute_summary", "week_summary"),
  ...
)
}
\arguments{
\item{traitData}{data frame with trait data}

\item{traitSignal}{data from with trait signals}

\item{traitnames}{names of `dataset: trait` without `timecol` information}

\item{response}{character string for type of response}

\item{timecol}{column to use for time}

\item{...}{additional parameters ignored}
}
\description{
Strains over Time
}
