% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainplot.R
\name{strainplot}
\alias{strainplot}
\title{GGplot by strain and condition}
\usage{
strainplot(
  datatraits,
  facet_strain = FALSE,
  condition = "sex_diet",
  shape_sex = FALSE,
  boxplot = FALSE,
  horizontal = FALSE
)
}
\arguments{
\item{datatraits}{data frame to be plotted}

\item{facet_strain}{facet by strain if `TRUE`}

\item{condition}{name of column with condition}

\item{shape_sex}{use different shape by sex if `TRUE`}

\item{boxplot}{overlay boxplot if `TRUE`}

\item{horizontal}{flip vertical and horizontal axis if `TRUE`}
}
\value{
object of class ggplot
}
\description{
GGplot by strain and condition
}
