% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainstats.R
\name{strainstats}
\alias{strainstats}
\title{Use Broom to Find Stats for Model Summaries}
\usage{
strainstats(
  object,
  signal = ifelse(is_condition, "strain * sex * condition", "strain * sex"),
  rest = ifelse(is_condition, "strain * sex + sex * condition", "sex"),
  calc_sd = TRUE
)
}
\arguments{
\item{object}{data frame in long format with trait data}

\item{signal}{signal factor combination as string for `formula`}

\item{rest}{rest factor combination as string for `formula`}

\item{calc_sd}{calculate SDs by `term` if `TRUE` (default)}
}
\value{
data frame with summaries by trait
}
\description{
Use Broom to Find Stats for Model Summaries
}
\examples{
strainstats(sampleData)
}
