% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_cor.R
\name{subset_eigen_cor}
\alias{subset_eigen_cor}
\title{Subset of eigen_cor object}
\usage{
subset_eigen_cor(
  x,
  facetname,
  colorname,
  facetmodules = flev,
  colormodules = clev,
  ...
)
}
\arguments{
\item{x}{of class `eigen_cor`}

\item{facetname}{facet name}

\item{colorname}{color name}

\item{facetmodules}{names of color modules to keep}

\item{colormodules}{names of color modules to keep}
}
\value{
data frame with colorname, facetname, correlation
}
\description{
Subset of eigen_cor object
}
