% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_traitPairs.R
\name{traitPairs}
\alias{traitPairs}
\title{GGplot of pairs of traits}
\usage{
traitPairs(
  traitData,
  traitnames,
  pair = trait_pairs(traitnames, sep),
  response = c("value", "mean", "signal"),
  sep = " ON ",
  ...
)
}
\arguments{
\item{traitData}{}

\item{traitnames}{}

\item{pair}{}

\item{sep}{}

\item{...}{}
}
\description{
GGplot of pairs of traits
}
