% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitPairs.R
\name{traitPairs}
\alias{traitPairs}
\alias{autoplot.traitPairs}
\alias{plot.traitPairs}
\title{Prepare pairs of traits for plotting}
\usage{
traitPairs(
  object,
  traitnames = attr(object, "traitnames"),
  pair = paste(traitnames[1:2], collapse = sep),
  sep = " ON ",
  ...
)

\method{autoplot}{traitPairs}(object, ...)

\method{plot}{traitPairs}(x, ...)
}
\arguments{
\item{object}{object of class `traitSolos`}

\item{traitnames}{trait names as `dataset: trait`}

\item{pair}{vector of trait name pairs, each joined by `sep`}

\item{sep}{pair separator}

\item{...}{ignored}
}
\value{
object of class `traitPairs`
}
\description{
Prepare pairs of traits for plotting
}
\examples{
out <- traitSolos(sampleData)
out2 <- traitPairs(out)
plot(out2, parallel_lines = TRUE)
}
