% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitpairs.R
\name{traitpairs}
\alias{traitpairs}
\title{Create trait pairs from trait names}
\usage{
traitpairs(traitnames, sep = " ON ")
}
\arguments{
\item{traitnames}{}

\item{sep}{string to separate trait names in pairs}
}
\description{
Create trait pairs from trait names
}
