% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_traitSolos.R, R/traitSolos.R
\name{ggplot_traitSolos}
\alias{ggplot_traitSolos}
\alias{traitSolos}
\alias{summary_traitSolos}
\alias{summary.traitSolos}
\title{GGplot by strain and condition}
\usage{
ggplot_traitSolos(object, ...)

traitSolos(
  traitData,
  traitSignal = partition(traitData),
  traitnames = unique(unite_datatraits(traitSignal)),
  response = c("value", "cellmean", "signal"),
  strains = names(foundr::CCcolors),
  abbrev = FALSE,
  sep = ": "
)

summary_traitSolos(object, customSettings = NULL, ...)

\method{summary}{traitSolos}(object, ...)
}
\arguments{
\item{object, x}{data frame to be plotted}

\item{...}{additional parameters}

\item{traitData}{data frame}

\item{traitSignal}{data frame}

\item{traitnames}{names of `dataset: trait` combinations to subset}

\item{response}{name of response to return}

\item{strains}{names of strains to subset}

\item{abbrev}{abbreviate names if `TRUE`}

\item{customSettings}{list of custom settings (including "condition_name" and "dataset")}

\item{facet_strain}{facet by strain if `TRUE`}

\item{shape_sex}{use different shape by sex if `TRUE`}

\item{boxplot}{overlay boxplot if `TRUE`}

\item{horizontal}{flip vertical and horizontal axis if `TRUE`}

\item{object}{object of class `traitSolos`}
}
\value{
object of class ggplot

data frame

summary table

summary table
}
\description{
GGplot by strain and condition

Select Traits and Strains for Response

Summary of traitSolos object

Summary of traitSolos object
}
\examples{
ggplot_traitSolos(traitSolos(sampleData))

out <- traitSolos(sampleData)
summary(out)
plot(out)
}
\seealso{
\code{\link{ggplot_traitSolos}}
}
