% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{One-sided Volcano Plot}
\usage{
volcano(
  object,
  termname = terms[1],
  threshold = c(SD = 1, p = 0.01),
  interact = FALSE,
  traitnames = TRUE
)
}
\arguments{
\item{object}{data frame from `strainstats`}

\item{termname}{name of `term` to show}

\item{threshold}{named vector for `SD` and `p.value`}

\item{interact}{prepare for interactive if `TRUE`}

\item{traitnames}{include trait names if `TRUE`}
}
\value{
ggplot object
}
\description{
One-sided Volcano Plot
}
\examples{
sampleStats <- strainstats(sampleData)
volcano(sampleStats)
}
