% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{Volcano Plot}
\usage{
volcano(
  object,
  termname = terms[1],
  threshold = threshold_default,
  interact = FALSE,
  traitnames = TRUE,
  facet = FALSE,
  ordername = c("p.value", "kME", "module", "size"),
  xlab = xlab_default,
  ylab = ylab_default,
  ...
)
}
\arguments{
\item{object}{data frame from `strainstats`}

\item{termname}{name of `term` to show}

\item{threshold}{named vector for `SD` and `p.value`}

\item{interact}{prepare for interactive if `TRUE`}

\item{traitnames}{include trait names if `TRUE`}

\item{facet}{facet on `strain` if `TRUE`}

\item{ordername}{name of column for Y ordering.}

\item{xlab, ylab}{axis labels}

\item{...}{additional parameters ignored}
}
\value{
ggplot object
}
\description{
Volcano Plot
}
\examples{
sampleStats <- strainstats(sampleData)
volcano(sampleStats)
}
