% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano.R
\name{volcano}
\alias{volcano}
\title{One-sided Volcano Plot}
\usage{
volcano(
  object,
  termname = terms[1],
  threshold = c(SD = 1, p = 0.01),
  interact = FALSE
)
}
\arguments{
\item{object}{data frame with `trait` names, `term` names, `SD` and `p.value`}

\item{termname}{name of `term` to show}

\item{threshold}{named vector for `SD` and `p.value`}

\item{interact}{prepare for interactive if `TRUE`}
}
\description{
One-sided Volcano Plot
}
